chore: extract engine tests to separate file (#16671)
